Extract typed session projections
