+++ server.ts
@@ -18,7 +18,6 @@ import {
   type ExtensionUIDialogOptions,
   type ExtensionUIContext,
   type SessionStartEvent,
-  type SlashCommandInfo,
 } from "@earendil-works/pi-coding-agent";
 import { createMockHarness } from "./server/mock.js";
 import { resolveBundledExtensionPaths, resolvePiWebExtensionPaths } from "./server/extensions.js";
@@ -29,6 +28,21 @@ import { assertDirectory, createDirectory, listDirectories } from "./server/shar
 import { gitCommitDetails, gitCwdFromRepoParam, gitDiff, gitLog, gitStatus, gitSync, isGitRepo, listGitRepos, readGitImage } from "./server/shared/git.js";
 import type { PiWebFooter, PiWebGitTab, PiWebHeaderAction, PiWebUi } from "./src/extensions.js";
 import type { PiWebSession } from "./server/types.js";
+import type { SlashCommandDto } from "./server/session/dto.js";
+import {
+  conversationTreeForSession,
+  getSessionSlashCommands,
+  isAssistantAbortedMessage,
+  isAssistantFailureMessage,
+  isIncompleteToolResultMessage,
+  messageEntryRefs,
+  projectSessionState,
+  sessionIsRetrying,
+  sessionStats,
+  simplifyMessage,
+  simplifyModel,
+  textFromContent,
+} from "./server/session/projection.js";
 
 const appDir = resolve(fileURLToPath(new URL(".", import.meta.url)));
 const distDir = join(appDir, "dist");
@@ -47,7 +61,7 @@ const noSession = process.env.PI_WEB_NO_SESSION === "1";
 const mockMode = process.env.PI_WEB_MOCK === "1";
 const execFileAsync = promisify(execFile);
 
-type WebSlashCommandInfo = Omit<SlashCommandInfo, "source"> & { source: SlashCommandInfo["source"] | "web" };
+type WebSlashCommandInfo = SlashCommandDto;
 
 const webSlashCommands: WebSlashCommandInfo[] = [
   { name: "help", description: "Show slash command help", source: "web", sourceInfo: { path: "<pi-web>", source: "pi-web", scope: "temporary", origin: "top-level" } },
@@ -146,31 +160,6 @@ function serveStatic(req: IncomingMessage, res: ServerResponse) {
 
   res.writeHead(200, { "content-type": contentTypes[extname(file)] || "application/octet-stream" });
   pipeReadStream(res, file);
-}
-
-function textFromContent(content: unknown): string {
-  if (typeof content === "string") return content;
-  if (!Array.isArray(content)) return "";
-  return content.map((part) => {
-    if (!part || typeof part !== "object") return "";
-    const p = part as Record<string, unknown>;
-    if (p.type === "text" && typeof p.text === "string") return p.text;
-    if (p.type === "image") return "[image]";
-    // toolCall parts are rendered as tool cards in the UI — omit from text
-    return "";
-  }).filter(Boolean).join("\n");
-}
-
-function simplifyModel(model: any) {
-  if (!model) return undefined;
-  return {
-    provider: model.provider,
-    id: model.id,
-    name: model.name || model.id,
-    reasoning: Boolean(model.reasoning),
-    contextWindow: model.contextWindow,
-    maxTokens: model.maxTokens,
-  };
 }
 
 function hasUserMessages(value: PiWebSession) {
@@ -279,348 +268,6 @@ function contentWithToolStartedAts(content: unknown, sessionFile?: string) {
   });
 }
 
-function appendMessageEntryRef(refs: Array<{ entryId?: string }>, entry: any) {
-  if (!entry || typeof entry !== "object") return;
-  if (entry.type === "message" || entry.type === "custom_message" || entry.type === "branch_summary" && entry.summary) {
-    const entryId = typeof entry.id === "string" && entry.id.trim() ? entry.id : undefined;
-    refs.push({ entryId });
-  }
-}
-
-function messageEntryRefs(targetSession: PiWebSession): Array<{ entryId?: string }> {
-  const getBranch = targetSession.sessionManager?.getBranch;
-  if (typeof getBranch !== "function") return [];
-
-  let branch: any[];
-  try {
-    branch = getBranch.call(targetSession.sessionManager);
-  } catch {
-    return [];
-  }
-  if (!Array.isArray(branch)) return [];
-
-  const refs: Array<{ entryId?: string }> = [];
-  let compaction: any | undefined;
-  for (const entry of branch) {
-    if (entry?.type === "compaction") compaction = entry;
-  }
-
-  if (!compaction) {
-    for (const entry of branch) appendMessageEntryRef(refs, entry);
-    return refs;
-  }
-
-  const compactionId = typeof compaction.id === "string" && compaction.id.trim() ? compaction.id : undefined;
-  refs.push({ entryId: compactionId });
-  const compactionIndex = branch.findIndex((entry) => entry?.type === "compaction" && entry?.id === compaction.id);
-  let foundFirstKept = false;
-  for (let index = 0; index < compactionIndex; index += 1) {
-    const entry = branch[index];
-    if (entry?.id === compaction.firstKeptEntryId) foundFirstKept = true;
-    if (foundFirstKept) appendMessageEntryRef(refs, entry);
-  }
-  for (let index = compactionIndex + 1; index < branch.length; index += 1) appendMessageEntryRef(refs, branch[index]);
-  return refs;
-}
-
-function simplifyMessage(message: unknown, toolCallArgs?: Map<string, Record<string, unknown>>, sessionFile?: string, entryId?: string) {
-  if (!message || typeof message !== "object") return message;
-  const m = message as Record<string, unknown>;
-  const content = contentWithToolStartedAts(m.content, sessionFile);
-  const entry = entryId ? { entryId } : {};
-  if (m.role === "bashExecution") {
-    return {
-      ...entry,
-      role: "bashExecution",
-      command: m.command,
-      output: m.output,
-      exitCode: m.exitCode,
-      cancelled: Boolean(m.cancelled),
-      truncated: Boolean(m.truncated),
-      fullOutputPath: m.fullOutputPath,
-      excludeFromContext: Boolean(m.excludeFromContext),
-      timestamp: m.timestamp,
-      raw: m,
-    };
-  }
-  if (m.role === "toolResult") {
-    const args = toolCallArgs?.get(m.toolCallId as string);
-    return {
-      ...entry,
-      role: "toolResult",
-      toolCallId: m.toolCallId,
-      toolName: m.toolName,
-      toolArgs: args,
-      isError: Boolean(m.isError),
-      text: textFromContent(m.content),
-      timestamp: m.timestamp,
-      raw: m,
-    };
-  }
-  const text = textFromContent(content);
-  const errorText = m.role === "assistant" && m.errorMessage ? assistantErrorPreview(m) : "";
-  const stopReasonText = m.role === "assistant" && !errorText ? assistantStopReasonPreview(m) : "";
-  const displayText = errorText || (text && stopReasonText ? `${text}\n\n${stopReasonText}` : stopReasonText || text);
-  const toolCalls = m.role === "assistant" && Array.isArray(content)
-    ? content.filter((part: any) => part?.type === "toolCall").map((part: any) => ({
-      id: part.id,
-      toolName: part.toolName || part.name || "tool",
-      args: part.arguments || part.args || {},
-      startedAt: part.startedAt,
-    }))
-    : undefined;
-  return {
-    ...entry,
-    role: m.role,
-    text: displayText,
-    toolCalls,
-    isError: Boolean(m.errorMessage || m.stopReason === "error" || stopReasonText),
-    timestamp: m.timestamp,
-    raw: content === m.content ? m : { ...m, content },
-  };
-}
-
-function truncatePreview(value: string, max = 220) {
-  const text = value.replace(/\s+/g, " ").trim();
-  return text.length > max ? `${text.slice(0, max - 1)}…` : text;
-}
-
-function entryMessage(entry: any) {
-  if (entry?.type === "message") return entry.message;
-  if (entry?.type === "custom_message") return { role: "custom", content: entry.content, timestamp: entry.timestamp };
-  return undefined;
-}
-
-function messageToolCalls(message: any) {
-  return Array.isArray(message?.content)
-    ? message.content.filter((part: any) => part?.type === "toolCall")
-    : [];
-}
-
-function toolCallName(part: any) {
-  return String(part?.toolName || part?.name || "tool");
-}
-
-function toolCallArgs(part: any) {
-  const args = part?.arguments || part?.args;
-  return args && typeof args === "object" ? args as Record<string, unknown> : {};
-}
-
-function shortArg(value: unknown, max = 90) {
-  const text = typeof value === "string" ? value : JSON.stringify(value ?? "");
-  return text.length > max ? `${text.slice(0, max - 1)}…` : text;
-}
-
-function toolCallPreview(part: any) {
-  const name = toolCallName(part);
-  const args = toolCallArgs(part);
-  if (name === "bash" && typeof args.command === "string") return `Tool call: bash ${shortArg(args.command, 120)}`;
-  if (typeof args.path === "string") return `Tool call: ${name} ${shortArg(args.path, 120)}`;
-  if (typeof args.query === "string") return `Tool call: ${name} ${shortArg(args.query, 120)}`;
-  if (typeof args.pattern === "string") return `Tool call: ${name} ${shortArg(args.pattern, 120)}`;
-  const first = Object.entries(args).find(([, value]) => typeof value === "string" || typeof value === "number" || typeof value === "boolean");
-  return first ? `Tool call: ${name} ${first[0]}=${shortArg(first[1], 90)}` : `Tool call: ${name}`;
-}
-
-function toolCallsPreview(message: any) {
-  const calls = messageToolCalls(message);
-  if (calls.length === 0) return "";
-  const [first] = calls;
-  const suffix = calls.length > 1 ? ` + ${calls.length - 1} more` : "";
-  return `${toolCallPreview(first)}${suffix}`;
-}
-
-function messageTextPreview(message: any) {
-  return textFromContent(message?.content || "");
-}
-
-const assistantHttpErrorLabels: Record<string, string> = {
-  "429": "Throttling error",
-  "500": "Server error",
-  "502": "Bad gateway",
-  "503": "Service unavailable",
-  "504": "Gateway timeout",
-  "529": "Overloaded",
-};
-
-function isAssistantHttpErrorStatus(code: string) {
-  return code in assistantHttpErrorLabels || /^[45]\d\d$/.test(code);
-}
-
-function assistantStatusLabel(label: string | undefined, code: string) {
-  const clean = (label || "").replace(/[_-]+/g, " ").replace(/\s+/g, " ").trim();
-  if (!clean || /^(?:http|status|error|request failed|model request failed)$/i.test(clean)) return assistantHttpErrorLabels[code] || `HTTP ${code}`;
-  return clean;
-}
-
-function assistantStatusErrorPreview(text: string) {
-  const labelled = text.match(/^([A-Za-z][A-Za-z0-9 _/-]*?):\s*(\d{3})(?=$|[\s:,-])/);
-  if (labelled && isAssistantHttpErrorStatus(labelled[2])) return `${assistantStatusLabel(labelled[1], labelled[2])} (${labelled[2]})`;
-  const leading = text.match(/^(?:HTTP\s*)?(\d{3})(?=$|[\s:,-])/i);
-  if (leading && isAssistantHttpErrorStatus(leading[1])) return `${assistantStatusLabel(undefined, leading[1])} (${leading[1]})`;
-  const generic = text.match(/^(Error|Request failed|Model request failed)\s*:?\s*(\d{3})(?=$|[\s:,-])/i);
-  if (generic && isAssistantHttpErrorStatus(generic[2])) return `${assistantStatusLabel(generic[1], generic[2])} (${generic[2]})`;
-  return "";
-}
-
-function assistantParsedErrorDetail(parsed: any) {
-  if (typeof parsed === "string") return parsed.trim();
-  if (!parsed || typeof parsed !== "object") return "";
-  if (parsed.error && typeof parsed.error === "object") return parsed.error.message || parsed.error.type || "";
-  return parsed.message || parsed.detail || parsed.error_description || "";
-}
-
-function assistantJsonErrorPreview(text: string) {
-  const trimmed = text.trim();
-  if (!((trimmed.startsWith("{") && trimmed.endsWith("}")) || (trimmed.startsWith("[") && trimmed.endsWith("]")))) return "";
-  try {
-    const detail = assistantParsedErrorDetail(JSON.parse(trimmed));
-    return detail ? `Error: ${detail}` : "";
-  } catch {
-    return "";
-  }
-}
-
-function assistantErrorPreview(message: any) {
-  const raw = String(message?.errorMessage || "").trim();
-  if (!raw) return "";
-  const jsonText = raw.replace(/^Codex error:\s*/i, "").trim();
-  return assistantJsonErrorPreview(jsonText)
-    || assistantStatusErrorPreview(jsonText)
-    || assistantStatusErrorPreview(raw)
-    || (raw.length > 180 ? `${raw.slice(0, 179)}…` : raw);
-}
-
-function assistantStopReasonPreview(message: any) {
-  const reason = String(message?.stopReason || "").trim();
-  if (!reason || reason === "stop" || reason === "toolUse") return "";
-  if (reason === "length") return "Response stopped because the model hit its output length limit.";
-  if (reason === "aborted") return "Response was aborted.";
-  return `Response stopped unexpectedly: ${reason}`;
-}
-
-function entryRole(entry: any) {
-  const message = entryMessage(entry);
-  if (message?.role === "assistant" && !messageTextPreview(message).trim()) {
-    if (messageToolCalls(message).length > 0) return "toolCall";
-    if (message.errorMessage || assistantStopReasonPreview(message)) return "error";
-  }
-  if (message?.role) return String(message.role);
-  switch (entry?.type) {
-    case "branch_summary": return "branchSummary";
-    case "compaction": return "compaction";
-    case "model_change": return "model";
-    case "thinking_level_change": return "thinking";
-    case "session_info": return "session";
-    case "label": return "label";
-    case "custom": return "custom";
-    default: return String(entry?.type || "entry");
-  }
-}
-
-function entryPreview(entry: any) {
-  const message = entryMessage(entry);
-  if (message) {
-    if (message.role === "toolResult") {
-      const text = textFromContent(message.content);
-      return `Tool result: ${message.toolName || "tool"}${text ? ` — ${text}` : ""}`;
-    }
-    const text = messageTextPreview(message);
-    if (text.trim()) return text;
-    const calls = toolCallsPreview(message);
-    if (calls) return calls;
-    const error = assistantErrorPreview(message);
-    if (error) return error;
-    const stopReason = assistantStopReasonPreview(message);
-    if (stopReason) return stopReason;
-    return message.role === "assistant" ? "Empty assistant message" : `${message.role || "Message"} message`;
-  }
-  switch (entry?.type) {
-    case "branch_summary": return entry.summary || "Branch summary";
-    case "compaction": return entry.summary || "Compaction summary";
-    case "model_change": return `Model changed to ${entry.provider || "provider"}/${entry.modelId || "model"}`;
-    case "thinking_level_change": return `Thinking level changed to ${entry.thinkingLevel || "unknown"}`;
-    case "session_info": return entry.name ? `Session named ${entry.name}` : "Session name cleared";
-    case "label": return entry.label ? `Label ${entry.targetId || "entry"} as ${entry.label}` : `Clear label on ${entry.targetId || "entry"}`;
-    case "custom": return `Custom entry${entry.customType ? `: ${entry.customType}` : ""}`;
-    default: return String(entry?.type || "Entry");
-  }
-}
-
-function countTreeNodes(nodes: any[]): number {
-  let count = 0;
-  const stack = [...nodes];
-  while (stack.length > 0) {
-    const node = stack.pop();
-    count += 1;
-    const children = Array.isArray(node?.children) ? node.children : [];
-    for (const child of children) stack.push(child);
-  }
-  return count;
-}
-
-function countBranchPoints(nodes: any[]): number {
-  let count = 0;
-  const stack = [...nodes];
-  while (stack.length > 0) {
-    const node = stack.pop();
-    const children = Array.isArray(node?.children) ? node.children : [];
-    if (children.length > 1) count += 1;
-    for (const child of children) stack.push(child);
-  }
-  return count;
-}
-
-function simpleTreeNode(node: any, activePathIds: Set<string>, leafId: string | null, childCount: number): any {
-  const entry = node?.entry || node;
-  const id = String(entry?.id || "");
-  return {
-    id,
-    parentId: typeof entry?.parentId === "string" ? entry.parentId : null,
-    type: String(entry?.type || "entry"),
-    role: entryRole(entry),
-    preview: truncatePreview(entryPreview(entry)),
-    timestamp: String(entry?.timestamp || ""),
-    label: typeof node?.label === "string" ? node.label : undefined,
-    labelTimestamp: typeof node?.labelTimestamp === "string" ? node.labelTimestamp : undefined,
-    childCount,
-    isOnActivePath: activePathIds.has(id),
-    isCurrentLeaf: Boolean(leafId && id === leafId),
-    children: [],
-  };
-}
-
-function simplifyTreeNodesFlat(roots: any[], activePathIds: Set<string>, leafId: string | null): any[] {
-  const nodes: any[] = [];
-  const stack = [...roots].reverse();
-  while (stack.length > 0) {
-    const node = stack.pop();
-    const children = Array.isArray(node?.children) ? node.children : [];
-    nodes.push(simpleTreeNode(node, activePathIds, leafId, children.length));
-    for (let index = children.length - 1; index >= 0; index -= 1) stack.push(children[index]);
-  }
-  return nodes;
-}
-
-function conversationTreeForSession(targetSession: PiWebSession) {
-  const manager = targetSession.sessionManager;
-  if (typeof manager.getTree !== "function") throw new Error("Session tree is not available");
-  const leafId = typeof manager.getLeafId === "function" ? manager.getLeafId() : null;
-  const activePath = typeof manager.getBranch === "function" ? manager.getBranch() : [];
-  const activePathIds = new Set(activePath.map((entry: any) => String(entry?.id || "")).filter(Boolean));
-  const roots = manager.getTree();
-  const nodes = simplifyTreeNodesFlat(roots, activePathIds, leafId);
-  return {
-    ok: true,
-    sessionId: targetSession.sessionId,
-    leafId,
-    activePathIds: Array.from(activePathIds),
-    entryCount: nodes.length,
-    branchPointCount: nodes.filter((node: any) => node.childCount > 1).length,
-    nodes,
-  };
-}
-
 function sessionCwd(targetSession: PiWebSession | any = session) {
   return String(targetSession?.sessionManager?.getCwd?.() || targetSession?.cwd || piCwd);
 }
@@ -669,34 +316,6 @@ function clearRuntimeStartedAt(targetSession: any, sessionFile = sessionPathKey(
     delete targetSession.runtimeStartedAt;
     delete targetSession.runtimeLastActivityAt;
   }
-}
-
-function messageRole(message: any) {
-  return String(message?.role || message?.raw?.role || "");
-}
-
-function messageStopReason(message: any) {
-  return String(message?.stopReason || message?.raw?.stopReason || "");
-}
-
-function messageErrorText(message: any) {
-  return typeof message?.errorMessage === "string"
-    ? message.errorMessage
-    : typeof message?.raw?.errorMessage === "string"
-      ? message.raw.errorMessage
-      : "";
-}
-
-function isAssistantFailureMessage(message: any) {
-  return messageRole(message) === "assistant" && (messageStopReason(message) === "error" || Boolean(messageErrorText(message).trim()));
-}
-
-function isAssistantAbortedMessage(message: any) {
-  return messageRole(message) === "assistant" && messageStopReason(message) === "aborted";
-}
-
-function isIncompleteToolResultMessage(message: any) {
-  return messageRole(message) === "toolResult";
 }
 
 type RetrySessionTarget =
@@ -798,10 +417,6 @@ async function retrySessionFromFailure(targetSession: PiWebSession) {
     internal._systemPromptOverride = undefined;
     if (typeof internal._flushPendingBashMessages === "function") internal._flushPendingBashMessages();
   }
-}
-
-function sessionIsRetrying(live: PiWebSession | undefined) {
-  return Boolean((live as any)?.isRetrying);
 }
 
 function runtimeForPath(path: string, overrides: { isRetrying?: boolean } = {}) {
@@ -915,100 +530,19 @@ async function listSessionInfos(extraCwds: string[] = []) {
   return groups.flat().sort((a, b) => Date.parse(b.modified) - Date.parse(a.modified));
 }
 
-function finiteNumber(value: unknown) {
-  return typeof value === "number" && Number.isFinite(value) ? value : 0;
-}
-
-function sessionDisplayName(targetSession: PiWebSession) {
-  return targetSession.getSessionName?.()?.trim()
-    || targetSession.sessionName?.trim()
-    || targetSession.sessionManager.getSessionName?.()?.trim()
-    || undefined;
-}
-
-function liveSessionTitle(targetSession: PiWebSession) {
-  const name = sessionDisplayName(targetSession);
-  if (name) return name;
-
-  for (const message of targetSession.messages as any[]) {
-    const text = textFromContent(message?.content).trim();
-    if (message?.role === "user" && text) return truncatePreview(text, 80);
-  }
-  return "New session";
-}
-
-function sessionStats(targetSession: PiWebSession) {
-  let input = 0;
-  let output = 0;
-  let cacheRead = 0;
-  let cacheWrite = 0;
-  let cost = 0;
-  let userMessages = 0;
-  let assistantMessages = 0;
-  let toolResults = 0;
-
-  const branch = targetSession.sessionManager.getBranch?.();
-  const entries = Array.isArray(branch) && branch.length > 0
-    ? branch.map((entry: any) => entry?.message ?? entry)
-    : targetSession.messages;
-
-  for (const message of entries as any[]) {
-    if (!message || typeof message !== "object") continue;
-    if (message.role === "user") userMessages++;
-    if (message.role === "toolResult") toolResults++;
-    if (message.role !== "assistant") continue;
-    assistantMessages++;
-    const usage = message.usage || {};
-    input += finiteNumber(usage.input);
-    output += finiteNumber(usage.output);
-    cacheRead += finiteNumber(usage.cacheRead);
-    cacheWrite += finiteNumber(usage.cacheWrite);
-    const usageCost = usage.cost || {};
-    const totalCost = finiteNumber(usageCost.total);
-    cost += totalCost || finiteNumber(usageCost.input) + finiteNumber(usageCost.output) + finiteNumber(usageCost.cacheRead) + finiteNumber(usageCost.cacheWrite);
-  }
-
-  const contextUsage = targetSession.getContextUsage?.() || undefined;
-  return {
-    userMessages,
-    assistantMessages,
-    toolResults,
-    totalMessages: entries.length,
-    tokens: {
-      input,
-      output,
-      cacheRead,
-      cacheWrite,
-      total: input + output + cacheRead + cacheWrite,
-    },
-    cost,
-    contextUsage,
-  };
-}
-
 function currentState(targetSession: PiWebSession = session) {
-  const isRetrying = sessionIsRetrying(targetSession);
-  const isRunning = Boolean(targetSession.isStreaming || isRetrying || targetSession.isCompacting);
-  const runtime = runtimeForPath(targetSession.sessionFile);
+  const projected = projectSessionState(targetSession, sessionCwd(targetSession));
+  const { thinkingLevels: _thinkingLevels, ...base } = projected;
+  const isRunning = Boolean(projected.isStreaming || projected.isRetrying || projected.isCompacting);
   return {
-    cwd: sessionCwd(targetSession),
-    sessionFile: targetSession.sessionFile,
-    sessionId: targetSession.sessionId,
-    sessionName: sessionDisplayName(targetSession),
-    sessionTitle: liveSessionTitle(targetSession),
-    isStreaming: targetSession.isStreaming,
-    isRetrying,
-    isCompacting: Boolean(targetSession.isCompacting),
+    ...base,
     runtimeStartedAt: typeof (targetSession as any).runtimeStartedAt === "string"
       ? (targetSession as any).runtimeStartedAt
       : runtimeStartedAtForPath(targetSession.sessionFile, isRunning),
     runtimeLastActivityAt: typeof (targetSession as any).runtimeLastActivityAt === "string"
       ? (targetSession as any).runtimeLastActivityAt
       : runtimeLastActivityAtForPath(targetSession.sessionFile, isRunning),
-    runtime,
-    model: simplifyModel(targetSession.model),
-    thinkingLevel: targetSession.thinkingLevel,
-    stats: sessionStats(targetSession),
+    runtime: runtimeForPath(targetSession.sessionFile),
     webFooters: webFooterEntries(targetSession),
     webHeaderActions: webHeaderActionEntries(targetSession),
     webGitTabs: webGitTabEntries(targetSession),
@@ -1020,39 +554,6 @@ function currentStateWithThinkingLevels(targetSession: PiWebSession = session) {
     ...currentState(targetSession),
     thinkingLevels: targetSession.getAvailableThinkingLevels(),
   };
-}
-
-function getSessionSlashCommands(value: any): WebSlashCommandInfo[] {
-  const commands: WebSlashCommandInfo[] = [];
-
-  for (const command of value.extensionRunner?.getRegisteredCommands?.() || []) {
-    commands.push({
-      name: command.invocationName || command.name,
-      description: command.description,
-      source: "extension",
-      sourceInfo: command.sourceInfo,
-    });
-  }
-
-  for (const template of value.promptTemplates || value.resourceLoader?.getPrompts?.().prompts || []) {
-    commands.push({
-      name: template.name,
-      description: template.description,
-      source: "prompt",
-      sourceInfo: template.sourceInfo,
-    });
-  }
-
-  for (const skill of value.resourceLoader?.getSkills?.().skills || []) {
-    commands.push({
-      name: `skill:${skill.name}`,
-      description: skill.description,
-      source: "skill",
-      sourceInfo: skill.sourceInfo,
-    });
-  }
-
-  return commands.filter((command) => typeof command.name === "string" && command.name.length > 0);
 }
 
 function getSlashCommands(value: any = session): WebSlashCommandInfo[] {
@@ -2490,7 +1991,7 @@ const server = createServer(async (req, res) => {
           }
         }
         const refs = messageEntryRefs(targetSession);
-        return sendJson(res, 200, { ok: true, messages: msgs.map((m: unknown, index: number) => simplifyMessage(m, toolCallArgs, targetSession.sessionFile, refs[index]?.entryId)) });
+        return sendJson(res, 200, { ok: true, messages: msgs.map((m: unknown, index: number) => simplifyMessage(m, { toolCallArgs, decorateContent: (content) => contentWithToolStartedAts(content, targetSession.sessionFile), entryId: refs[index]?.entryId })) });
       }
 
       if (method === "GET" && url.pathname === "/api/sessions") {
